Move BulletedList, TexTextFromPresetString and Title to their own file
